requirements.txt
streamlit==1.38.0
torch==2.4.1
transformers==4.45.2
pillow==10.4.0
gtts==2.5.3
requests==2.32.3
accelerate==1.0.1
